Fix a bunch of warnings under gcc 4.5, especially a few where NULL was used in non-pointer contexts which generated a large amount of build noise.

+++ libmesh/plugins/common-filters/TriangulateFilter.cpp
@@ -96,7 +96,7 @@ FilterDataPtr TriangulateFilter::apply(FilterDataPtr input) {
                     // the different triplet for each winding order: 0, 1, 2 or
                     // 1, 0, 2.
 
-                    for(int tri_idx = 2; tri_idx < prim_it->indices.size(); tri_idx++) {
+                    for(int tri_idx = 2; tri_idx < (int)prim_it->indices.size(); tri_idx++) {
                         if (tri_idx % 2 == 0) {
                             new_prim.indices.push_back( prim_it->indices[tri_idx-2] );
                             new_prim.indices.push_back( prim_it->indices[tri_idx-1] );
@@ -119,7 +119,7 @@ FilterDataPtr TriangulateFilter::apply(FilterDataPtr input) {
                     // Triangle fans are pretty simple. They generate n
                     // triangles for n+2 vertices. Triangle n is formed by
                     // vertices 0, n+1, n+2, in that order.
-                    for(int tri_idx = 2; tri_idx < prim_it->indices.size(); tri_idx++) {
+                    for(int tri_idx = 2; tri_idx < (int)prim_it->indices.size(); tri_idx++) {
                         new_prim.indices.push_back( prim_it->indices[0] );
                         new_prim.indices.push_back( prim_it->indices[tri_idx-1] );
                         new_prim.indices.push_back( prim_it->indices[tri_idx] );
